backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pillow==10.1.0

# Deep Learning (uncomment when needed)
# torch==2.1.0
# transformers==4.35.0
# diffusers==0.24.0
# compel==0.1.10

# Optional utilities
requests==2.31.0
